chore(client): further clean up
