+++ app/client/src/components/VList/VList.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable */ // FIXME:
 import * as React from 'react';
 
-import useCache from '@local/hooks/useCache';
+import { useCache } from '@local/features/core';
 
 interface Props {
     children: React.ReactNode | React.ReactNodeArray;
