+++ app/client/src/features/accounts/UserMenu/UserMenu.tsx
@@ -26,7 +26,7 @@ import { graphql, usePreloadedQuery, PreloadedQuery } from 'react-relay';
 
 import { ResponsiveDialog } from '@local/components/ResponsiveDialog';
 import type { UserMenuQuery } from '@local/__generated__/UserMenuQuery.graphql';
-import { useIsClient } from '@local/hooks';
+import { useIsClient } from '@local/features/core';
 import { useUser } from '../useUser';
 import { LoginForm } from '../LoginForm';
 import { RegisterForm } from '../RegisterForm';
