+++ app/client/src/features/core/ThemeContext.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { ThemeProvider as MUIThemeProvider } from '@material-ui/core/styles';
 
-import themes, { Themes } from '@local/theme';
+import { themes, TThemes } from './theme';
 
 interface Props {
     children: JSX.Element | JSX.Element[];
@@ -10,7 +10,7 @@ interface Props {
 export const ThemeSelector = React.createContext<(() => void)[]>([() => {}]);
 
 export function ThemeProvider({ children }: Props) {
-    const [state, setState] = React.useState<keyof Themes>('dark');
+    const [state, setState] = React.useState<keyof TThemes>('dark');
     return (
         <ThemeSelector.Provider value={[() => setState(state === 'light' ? 'dark' : 'light')]}>
             <MUIThemeProvider theme={themes[state]}>{children}</MUIThemeProvider>
