+++ app/client/src/features/core/__mocks__/useIntersectionObserver.ts
