+++ app/client/src/features/core/__mocks__/useSnack.ts
@@ -1,0 +1,3 @@
+export default function useSnack() {
+    return { displaySnack: () => true, closeSnack: () => true };
+}
