+++ app/client/src/features/core/index.tsx
@@ -1,3 +1,10 @@
 export * from './SnackContext';
 export * from './ThemeContext';
 export * from './useSnack';
+export * from './relay';
+export * from './useConnection';
+export * from './useForm';
+export * from './useIntersectionObserver';
+export * from './useCache';
+export * from './useIsClient';
+export * from './theme';
