+++ app/client/src/features/core/useCache.ts
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 // TODO: could be extended to use session storage?
-export default function useCache<T>(initialValue: T) {
+export function useCache<T>(initialValue: T) {
     const ref = React.useRef(initialValue);
     return ref.current;
 }
