+++ app/client/src/features/core/useScrollTo.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import useIntersectionObserver from '@local/hooks/useIntersectionObserver';
+import { useIntersectionObserver } from './useIntersectionObserver';
 
 export default function useScrollTo(anchor: React.RefObject<HTMLElement>) {
     const [isAnchorInView, setIsAnchorInView] = React.useState<boolean>(false);
