+++ app/client/src/features/events/Questions/QuestionForm/QuestionForm.tsx
@@ -6,7 +6,7 @@ import { FormTitle } from '@local/components/FormTitle';
 import { FormContent } from '@local/components/FormContent';
 import { FormActions } from '@local/components/FormActions';
 import { TextField } from '@local/components/TextField';
-import { useForm } from '@local/hooks';
+import { useForm } from '@local/features/core';
 
 export type TQuestionFormState = { question: string };
