+++ app/client/src/features/organizations/OrgForm.tsx
@@ -2,7 +2,7 @@ import { Button } from '@material-ui/core';
 
 import { LoadingButton } from '@local/components/LoadingButton';
 import { TextField } from '@local/components/TextField';
-import { useForm } from '@local/hooks/useForm';
+import { useForm } from '@local/features/core';
 import type { CreateOrganization } from '@local/graphql-types';
 import { Form } from '@local/components/Form';
 import { FormContent } from '@local/components/FormContent';
