--- app/client/src/hooks/__mocks__/useSnack.ts
@@ -1,10 +1,0 @@
-export default function useSnack() {
-    return [
-        () => {
-            return true;
-        },
-        () => {
-            return false;
-        },
-    ];
-}
