+++ app/client/src/pages/_document.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Document, { Html, Head, Main, NextScript } from 'next/document';
 import { ServerStyleSheets } from '@material-ui/core/styles';
-import theme from '@local/theme';
+import { themes } from '@local/features/core';
 
 export default class MyDocument extends Document {
     render() {
@@ -9,7 +9,7 @@ export default class MyDocument extends Document {
             <Html lang='en'>
                 <Head>
                     {/* PWA primary color */}
-                    <meta name='theme-color' content={theme.dark.palette.primary.main} />
+                    <meta name='theme-color' content={themes.dark.palette.primary.main} />
                     <link
                         rel='stylesheet'
                         href='https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap'
@@ -59,7 +59,6 @@ MyDocument.getInitialProps = async (ctx) => {
         });
 
     const initialProps = await Document.getInitialProps(ctx);
-    
 
     return {
         ...initialProps,
