+++ app/client/src/reducers/index.ts
@@ -3,23 +3,17 @@ import { combineReducers, createStore } from '@reduxjs/toolkit';
 import { composeWithDevTools } from 'redux-devtools-extension';
 import { reducer as langReducer } from './language';
 
-import chatReducer from './chat';
 import feedReducer from './question-feed';
 import queueReducer from './quesiton-queue';
-import playlistReducer from './playlist';
 
 export default combineReducers({
     questions: feedReducer,
-    chat: chatReducer,
     queue: queueReducer,
-    playlist: playlistReducer,
     language: langReducer,
 });
 
-export * from './chat';
 export * from './quesiton-queue';
 export * from './question-feed';
-export * from './playlist';
 export * from './language';
 
 // function initStore(preloadedState = initialState) {
