--- app/client/src/utils/axios.ts
@@ -1,10 +1,0 @@
-// import axios from 'axios';
-
-// NOTE: remove this later
-// axios.defaults.headers = {
-//     post: {
-//         'Content-Type': 'application/json',
-//     },
-// };
-
-export default {};
